backend/requirements.txt
Django==4.2.13
djangorestframework==3.15.1
djangorestframework-simplejwt==5.3.1
django-cors-headers==4.3.1
django-environ==0.11.2
Pillow==10.3.0
psycopg2-binary==2.9.9
gunicorn==22.0.0
whitenoise==6.6.0

# Dev / Test
pytest==8.1.1
pytest-django==4.8.0
factory-boy==3.3.0
Faker==24.11.0
